src/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
